+++ packages/cmd/__tests__/CommandProcessor.test.ts
@@ -22,6 +22,7 @@ import { SharedOptions } from "../src/utils/SharedOptions";
 import { CommandProfileLoader } from "../src/profiles/CommandProfileLoader";
 import { CliUtils } from "../../utilities/src/CliUtils";
 import { WebHelpManager } from "../src/help/WebHelpManager";
+import { ImperativeConfig } from "../../utilities/src/ImperativeConfig";
 
 jest.mock("../src/syntax/SyntaxValidator");
 jest.mock("../src/utils/SharedOptions");
@@ -245,6 +246,7 @@ describe("Command Processor", () => {
     afterEach(() => {
         process.stdout.write = ORIGINAL_STDOUT_WRITE;
         process.stderr.write = ORIGINAL_STDERR_WRITE;
+        jest.restoreAllMocks();
     });
 
     it("should allow us to create an instance", () => {
@@ -1396,6 +1398,11 @@ describe("Command Processor", () => {
             };
         });
 
+        const mockConfigReload = jest.fn();
+        jest.spyOn(ImperativeConfig, "instance", "get").mockReturnValue({
+            config: { reload: mockConfigReload }
+        } as any);
+
         const parms: any = {
             arguments: {
                 _: ["check", "for", "banana"],
@@ -1408,6 +1415,7 @@ describe("Command Processor", () => {
         const commandResponse: ICommandResponse = await processor.invoke(parms);
         expect(commandResponse).toBeDefined();
         expect(commandResponse).toMatchSnapshot();
+        expect(mockConfigReload).toHaveBeenCalledTimes(1);
     });
 
     it("should extract arguments not specified on invoke from a profile and merge with args", async () => {