+++ packages/imperative/__tests__/ConfigurationLoader.test.ts
@@ -12,6 +12,12 @@
 import { ConfigurationLoader } from "..";
 import { IImperativeOverrides } from "../src/doc/IImperativeOverrides";
 import { IApimlSvcAttrs } from "../src/doc/IApimlSvcAttrs";
+import { homedir } from "os";
+import * as path from "path";
+
+function getCallerFile(file: string) {
+    return require(file);
+}
 
 describe("ConfigurationLoader", () => {
     describe("overrides", () => {
@@ -68,6 +74,57 @@ describe("ConfigurationLoader", () => {
             );
 
             expect(result.apimlConnLookup).toEqual(apimlConnLookup);
+        });
+    });
+
+    describe("ConfigurationModule", () => {
+        it("should return a config", () => {
+
+            const result = ConfigurationLoader.load(
+                {
+                    configurationModule: path.join(__dirname, "__model__", "Sample.configuration.ts")
+                },
+                {},
+                getCallerFile
+            );
+
+            expect(result.productDisplayName).toEqual("Zowe");
+            expect(result.rootCommandDescription).toEqual("Sample");
+            expect(result.envVariablePrefix).toEqual("Sample");
+            expect(result.defaultHome).toEqual(homedir());
+        });
+        it("should return a config without changing the name", () => {
+
+            const result = ConfigurationLoader.load(
+                {
+                    productDisplayName: "notZowe",
+                    configurationModule: path.join(__dirname, "__model__", "Sample.configuration.ts")
+                },
+                {},
+                getCallerFile
+            );
+
+            expect(result.productDisplayName).toEqual("Zowe");
+            expect(result.rootCommandDescription).toEqual("Sample");
+            expect(result.envVariablePrefix).toEqual("Sample");
+            expect(result.defaultHome).toEqual(homedir());
+        });
+        it("should return a config with daemonMode", () => {
+
+            const result = ConfigurationLoader.load(
+                {
+                    configurationModule: path.join(__dirname, "__model__", "Sample.configuration.ts"),
+                    daemonMode: true
+                },
+                {},
+                getCallerFile
+            );
+
+            expect(result.productDisplayName).toEqual("Zowe");
+            expect(result.rootCommandDescription).toEqual("Sample");
+            expect(result.envVariablePrefix).toEqual("Sample");
+            expect(result.defaultHome).toEqual(homedir());
+            expect(result.daemonMode).toEqual(true);
         });
     });
 });