+++ packages/imperative/src/config/cmd/init/init.handler.ts
@@ -15,7 +15,6 @@ import { Config, ConfigConstants, ConfigSchema, IConfig } from "../../../../../c
 import { IProfileProperty } from "../../../../../profiles";
 import { ConfigBuilder } from "../../../../../config/src/ConfigBuilder";
 import { IConfigBuilderOpts } from "../../../../../config/src/doc/IConfigBuilderOpts";
-import { CredentialManagerFactory } from "../../../../../security";
 import { coercePropValue, secureSaveError } from "../../../../../config/src/ConfigUtils";
 import { OverridesLoader } from "../../../OverridesLoader";
 import * as JSONC from "comment-json";
@@ -40,7 +39,7 @@ export default class InitHandler implements ICommandHandler {
         this.params = params;
 
         // Load the config and set the active layer according to user options
-        await this.ensureCredentialManagerLoaded();
+        await OverridesLoader.ensureCredentialManagerLoaded();
         const config = ImperativeConfig.instance.config;
         const configDir = params.arguments.globalConfig ? null : process.cwd();
         config.api.layers.activate(params.arguments.userConfig, params.arguments.globalConfig, configDir);
@@ -50,7 +49,7 @@ export default class InitHandler implements ICommandHandler {
         if (params.arguments.dryRun && params.arguments.dryRun === true) {
             let dryRun = await this.initForDryRun(config, params.arguments.userConfig);
 
-            if (params.arguments.prompt !== false && !CredentialManagerFactory.initialized && config.api.secure.secureFields().length > 0) {
+            if (params.arguments.prompt !== false && config.api.secure.loadFailed && config.api.secure.secureFields().length > 0) {
                 const warning = secureSaveError();
                 params.response.console.log(TextUtils.chalk.yellow("Warning:\n") +
                     `${warning.message} Skipped prompting for credentials.\n\n${warning.additionalDetails}\n`);
@@ -100,7 +99,7 @@ export default class InitHandler implements ICommandHandler {
         } else {
             await this.initWithSchema(config, params.arguments.userConfig, params.arguments.overwrite, params.arguments.forSure);
 
-            if (params.arguments.prompt !== false && !CredentialManagerFactory.initialized && config.api.secure.secureFields().length > 0) {
+            if (params.arguments.prompt !== false && config.api.secure.loadFailed && config.api.secure.secureFields().length > 0) {
                 const warning = secureSaveError();
                 params.response.console.log(TextUtils.chalk.yellow("Warning:\n") +
                     `${warning.message} Skipped prompting for credentials.\n\n${warning.additionalDetails}\n`);
@@ -109,17 +108,6 @@ export default class InitHandler implements ICommandHandler {
             // Write the active created/updated config layer
             await config.save(false);
             params.response.console.log(`Saved config template to ${layer.path}`);
-        }
-    }
-
-    /**
-     * If CredentialManager was not already loaded by Imperative.init, load it
-     * now before performing config operations in the init handler.
-     */
-    private async ensureCredentialManagerLoaded() {
-        if (!CredentialManagerFactory.initialized) {
-            await OverridesLoader.loadCredentialManager(ImperativeConfig.instance.loadedConfig,
-                ImperativeConfig.instance.callerPackageJson);
         }
     }
 
@@ -174,7 +162,7 @@ export default class InitHandler implements ICommandHandler {
      */
     private async promptForProp(propName: string, property: IProfileProperty): Promise<any> {
         // skip prompting in CI environment
-        if (this.params.arguments.prompt === false || !CredentialManagerFactory.initialized) {
+        if (this.params.arguments.prompt === false || ImperativeConfig.instance.config.api.secure.loadFailed) {
             return null;
         }